The Responsive GUI Quest - Part 2
After leaving home and finding info on taverns, I find out a guy named viewPort has access to the information of the screen's change in resolution. I convinced him to work with my partner, ScreenResize, and cam now changes to the correct size in response to it. Now, to deal with the mighty beast: GUI.

- Created ScreenResize interface that detects a change in screen's resolution.
- GUI is partially responsive now.