Chnaged GUI framework to Lemur (Part 1 - Changing libs)
** WHY?!? **

When I started learning how to build GUI's for jME, my first choice was Nifty, mainly for two reasons:

- It was the 1st one at that time
- It used XML to layout the GUI, which I thought it was awesome.

But, Nifty always gave me headaches and suffer when trying to learn it. There was no reasonable documentation, it's github wiki is half complete, I couldn't find help, it's manual is outdated, and etc... And, when I realized, I lost the purposes of using it:

- There were now more GUI frameworks (Lemur and tonegod), specifically for jME (Nifty is more global)
- I was designign half of the GUI in Java, not XML

So, I finally decided to stop it. I'm changing it to Lemur. I'm still unsure wheter I'll use Lemur or tonegod, but I'll figure that out later.

** WARNING **

This commit only deleted Nifty libs and adds Lemur. Thus, no changed to code were made, and therefore it's broken until I fix it. Don't try to build it untill I fix it in the next commit!